Update 1 to the interaction system
The player can now spawn tomato's by going to the spawner and pressing 'g'. pressing 'h', "drops" the inventory its not perfect as only would work for the tomato's rn. 'i' prints inventory in debugger.
I changed which node holds the vegetable spawner script to the rigidbody for my ease of use with the interaction system, can be changed back with some work.